+++ osu.Game.Rulesets.Mania/Mods/ManiaModMirror.cs
@@ -10,16 +10,8 @@ using osu.Game.Rulesets.Mania.Beatmaps;
 
 namespace osu.Game.Rulesets.Mania.Mods
 {
-    public class ManiaModMirror : Mod, IApplicableToBeatmap
+    public class ManiaModMirror : ModMirror, IApplicableToBeatmap
     {
-        public override string Name => "Mirror";
-        public override string Acronym => "MR";
-
-        public override string Description => "镜像模式";
-        public override ModType Type => ModType.Conversion;
-        public override double ScoreMultiplier => 1;
-        public override bool Ranked => true;
-
         public void ApplyToBeatmap(IBeatmap beatmap)
         {
             var availableColumns = ((ManiaBeatmap)beatmap).TotalColumns;